some commands were brought out the main file
