+++ bot/commands/__init__.py
@@ -1,0 +1,4 @@
+from .request import command_request
+from .avaliable import command_avaliable
+from .qa import command_qa
+__ALL__ = ['command_request', 'command_avaliable','qa']
